ansible_automation/requirements.txt
bravado==11.0.3
cryptography==42.0.3
pem==18.1.0
